Update Project - Part 1 Sentiment Classifier.ipynb
One test case fails in previous case in the online Runestone Interactive Environment, when the case is "what a truly Wonderful day it is today! #incredible".
The Expected output must be 2 (wonderful, incredible) but the actual output comes out to be only 1 (incredible).
So, the lower() method change the string in lowercase and then compares it to the positive words strings.